requirement.txt
accelerate==1.4.0
aiofiles==23.2.1
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
albucore==0.0.23
albumentations==2.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.13.2
anyio==4.8.0
arrow==1.3.0
async-timeout==5.0.1
attrs==25.1.0
av==14.1.0
binaryornot==0.4.4
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cookiecutter==2.6.0
cycler==0.12.1
datasets==3.3.0
decord==0.6.0
deepspeed==0.16.4
dill==0.3.8
distro==1.9.0
duckdb==1.2.0
einops==0.8.1
evaluate==0.4.3
exceptiongroup==1.2.2
fastapi==0.115.11
ffmpy==0.5.0
filelock==3.13.1
flash_attn==2.7.4.post1
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2024.6.1
gradio==5.20.0
gradio_client==1.7.2
groovy==0.1.2
h11==0.14.0
hf_transfer==0.1.9
hjson==3.1.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.1
idna==3.10
imageio==2.37.0
imgaug==0.4.0
iniconfig==2.1.0
Jinja2==3.1.4
jiter==0.9.0
joblib==1.4.2
kiwisolver==1.4.8
latex2sympy2_extended==1.0.9
lazy_loader==0.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
math-verify==0.6.0
matplotlib==3.10.0
mdurl==0.1.2
mpi4py==4.0.3
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.3
ninja==1.11.1.3
numpy==2.1.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.570.86
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.1.105
nvidia-nvtx-cu12==12.1.105
openai==1.66.2
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
orjson==3.10.15
packaging==24.2
pandas==2.2.3
peft==0.14.0
pillow==11.0.0
pluggy==1.5.0
propcache==0.3.0
psutil==7.0.0
py-cpuinfo==9.0.0
pyarrow==19.0.1
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.19.1
pyparsing==3.2.1
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
python-slugify==8.0.4
pytz==2025.1
PyYAML==6.0.2
qwen-vl-utils==0.0.8
regex==2024.11.6
requests==2.32.3
rich==13.9.4
ruff==0.9.9
safehttpx==0.1.6
safetensors==0.5.2
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
semantic-version==2.10.0
shapely==2.0.7
shellingham==1.5.4
simsimd==6.2.1
six==1.17.0
sniffio==1.3.1
starlette==0.46.0
stringzilla==3.11.3
sympy==1.13.1
text-unidecode==1.3
threadpoolctl==3.5.0
tifffile==2025.2.18
tokenizers==0.21.0
tomli==2.2.1
tomlkit==0.13.2
torch==2.3.0+cu121
torchaudio==2.3.0+cu121
torchvision==0.18.0+cu121
tqdm==4.67.1
transformers==4.49.0
triton==2.3.0
trl==0.15.1
typer==0.15.2
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
websockets==15.0
xxhash==3.5.0
yarl==1.18.3
